requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
openai==1.58.1
python-multipart==0.0.20
requests==2.31.0
beautifulsoup4==4.12.3
newspaper3k==0.2.8
lxml[html_clean]>=5.3.0
google-api-python-client==2.118.0
